Changed motor control classes to use float (single) instead of double in order to have System.Math functions available that do not throw a NotImplementedException, but instead use the single precision FPU of the Netduino 3 boards. NOTE: All samples compile without warnings or errors, but only is actually tested since this change.

+++ source/AbstractIO.AdafruitMotorShieldV2/Base/IServoControl.cs
@@ -27,13 +27,13 @@ namespace AbstractIO.AdafruitMotorShieldV2
         ///     rotation and 1.0 represents full counterclockwise rotation.
         /// </summary>
         /// <value>The position expressed as a fraction of unity, i.e. in the range 0.0 to +1.0 inclusive.</value>
-        double Position { get; set; }
+        float Position { get; set; }
         /// <summary>
         ///     Gets or sets the angular position (in positive degrees) of the servo motor. The zero-point is the position of maximum 
         ///     clockwise rotation and the angle is measured counter-clockwise.
         ///     The maximum angle is implementation dependent.
         /// </summary>
         /// <value>The angle, in positive degrees, away from full clockwise displacement.</value>
-        double Angle { get; set; }
+        float Angle { get; set; }
     }
 }
